iNNtrospect/requirements.txt
matplotlib==3.7.3
numpy==1.24.3
pandas==2.0.3
Pillow==10.2.0
scikit_learn==1.3.2
scipy==1.10.1
tensorflow==2.13.1
umap_learn==0.5.4
